add logger to target.rb, rename the helper
Add logger to target.rb, when require "yast2/execute"
Cheetah.default_options = { logger: Y2Logger.instance } is set,
so Cheetah will add logs to y2log when execute commands.

rename example_helper.rb -> iscsi-lio-server_helper.rb
