+++ src/lib/dialogs/UI_dialogs.rb
@@ -1,7 +1,4 @@
-# Simple example to demonstrate object API for CWM
-
-# require_relative "example_helper"
-require './src/lib/helps/example_helper.rb'
+require './src/lib/helps/iscsi-lio-server_helper.rb'
 require './src/lib/TargetData.rb'
 require 'cwm/widget'
 require 'ui/service_status'
